+++ src/monitor/node_metadata.h
@@ -100,9 +100,9 @@ typedef enum SyncState
  * AutoFailoverNode, for consistency. Well, apart when registering, where we
  * don't have the node id and/or the node name yet.
  */
-#define NODE_FORMAT "node %ld \"%s\" (%s:%d)"
+#define NODE_FORMAT "node %lld \"%s\" (%s:%d)"
 #define NODE_FORMAT_ARGS(node) \
-	node->nodeId, node->nodeName, node->nodeHost, node->nodePort
+	(long long) node->nodeId, node->nodeName, node->nodeHost, node->nodePort
 
 /*
  * AutoFailoverNode represents a Postgres node that is being tracked by the
