+++ src/monitor/notifications.c
@@ -84,7 +84,7 @@ NotifyStateChange(AutoFailoverNode *node, char *description)
 	escape_json(payload, node->formationId);
 
 	appendStringInfo(payload, ", \"groupId\": %d", node->groupId);
-	appendStringInfo(payload, ", \"nodeId\": %ld", node->nodeId);
+	appendStringInfo(payload, ", \"nodeId\": %lld", (long long) node->nodeId);
 
 	appendStringInfo(payload, ", \"name\": ");
 	escape_json(payload, node->nodeName);
